port functions to analysis
